refactor: state directory store로 이름 변경 - 상태들을 저장하는 공간으로 store라는 디렉토리명이 더 어울린다고 판단
