+++ src/api/project/searchMember.ts
@@ -1,5 +1,5 @@
 import { authAxios } from '@utils/axios';
-import { User } from '@state/type';
+import { User } from '@store/type';
 
 const searchMember = async (searchQuery: string): Promise<User[]> => {
   const result = await authAxios.get(`/user/${searchQuery}`);
