+++ src/app.tsx
@@ -5,7 +5,7 @@ import { createStore, applyMiddleware } from 'redux';
 import { Provider } from 'react-redux';
 import thunkMiddleware from 'redux-thunk';
 import { composeWithDevTools } from 'redux-devtools-extension';
-import rootReducer from '@state/store';
+import rootReducer from '@store/store';
 import Router from './route';
 
 const store = createStore(
