+++ src/pages/Issue/components/Events/EventsTab.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
 import { Paper } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
-import { SearchEvent, Event } from '@state/type';
+import { SearchEvent, Event } from '@store/type';
 import SearchBar from './SearchBar';
 import EventsTable from './EventsTable';
