+++ src/pages/Issue/components/Events/EventsTable.tsx
@@ -10,7 +10,7 @@ import {
   Paper,
 } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
-import { Event } from '@state/type';
+import { Event } from '@store/type';
 
 interface Props {
   events: Event[];
