+++ src/pages/Issue/components/IssueTabs.tsx
@@ -2,7 +2,7 @@ import React, { FC, useState } from 'react';
 import { Link } from 'react-router-dom';
 import { makeStyles } from '@material-ui/core/styles';
 import { Paper, Tabs, Tab } from '@material-ui/core';
-import { Issue } from '@state/type';
+import { Issue } from '@store/type';
 import DetailsTabContainer from './Details/DetailsTabContainer';
 import EventsTabContainer from './Events/EventsTabContainer';
 import IssueTagsContainer from './IssueTagsContainer';
