+++ src/pages/Issue/components/IssueTags.tsx
@@ -2,7 +2,7 @@ import React, { FC, Dispatch, SetStateAction } from 'react';
 import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
 import Paper from '@material-ui/core/Paper';
 import Grid from '@material-ui/core/Grid';
-import { SearchResult } from '@state/type';
+import { SearchResult } from '@store/type';
 import TagBarChart from './TagBarChart';
 
 const useStyles = makeStyles((theme: Theme) =>
