+++ src/pages/Issue/components/TagBarChart.tsx
@@ -3,7 +3,7 @@ import BillboardChart from 'react-billboardjs';
 import Button from '@material-ui/core/Button';
 import styled from '@emotion/styled';
 import 'react-billboardjs/lib/billboard.css';
-import { SearchResult } from '@state/type';
+import { SearchResult } from '@store/type';
 
 const GridHeader = styled.div`
   display: flex;
