+++ src/pages/IssueList/IssueList.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import { CssBaseline, Grid, Container } from '@material-ui/core';
-import { Docs, Issue } from '@state/type';
+import { Docs, Issue } from '@store/type';
 import AppbarShift from '../layout/appbarshift';
-import IssueTable from './components/issuetable';
+import IssueTable from './components/IssueTable';
 
 interface Props {
   issues: Docs<Issue>;
