+++ src/pages/IssueList/IssueListContainer.tsx
@@ -3,7 +3,7 @@ import { useSelector, DefaultRootState } from 'react-redux';
 import getIssues from '@api/issue/getIssues';
 import AlertDialog from '@common/AlertDialog';
 
-import { Issue, Docs } from '@state/type';
+import { Issue, Docs } from '@store/type';
 import IssueList from './IssueList';
 
 interface State extends DefaultRootState {
