+++ src/pages/ProjectList/ProjectList.tsx
@@ -6,9 +6,9 @@ import { CssBaseline, Grid, Container, Tooltip, Fab } from '@material-ui/core';
 import { Pagination } from '@material-ui/lab';
 import { Add } from '@material-ui/icons';
 
-import { Project, Docs } from '@state/type';
+import { Project, Docs } from '@store/type';
 
-import SimpleCard from './components/simplecard';
+import SimpleCard from './components/SimpleCard';
 import AppbarShift from '../layout/appbarshift';
 
 const useStyles = makeStyles((theme) => ({
