+++ src/pages/layout/appbarshift.tsx
@@ -11,7 +11,7 @@ import {
 } from '@material-ui/core';
 import { Menu, ChevronLeft, LockOpen } from '@material-ui/icons';
 import SecondLogo from '@common/SecondLogo';
-import MainListItems from './sidebar';
+import MainListItems from './SideBar';
 
 const drawerWidth = 200;
