+++ src/store/curProject/curProjectReducer.ts
