+++ src/store/projects/projectsActions.ts
@@ -1,4 +1,4 @@
-import { Project, Docs } from '@state/type';
+import { Project, Docs } from '@store/type';
 
 export const GET_PROJECTS_REQUEST = 'GET_PROJECTS_REQUEST';
 export const GET_PROJECTS_SUCCESS = 'GET_PROJECTS_SUCCESS';
