+++ tsconfig.json
@@ -18,7 +18,7 @@
       "@common/*": ["src/common/*"],
       "@hooks/*": ["src/hooks/*"],
       "@pages/*": ["src/pages/*"],
-      "@state/*": ["src/state/*"],
+      "@store/*": ["src/store/*"],
       "@themes/*": ["src/themes/*"],
       "@utils/*": ["src/utils/*"],
       "@public/*":["public/*"]
